WSTEAM1-313: Move Uploader component into Embeds dir
