+++ src/app/components/Embeds/Uploader/index.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen } from '../react-testing-library-with-providers';
+import { render, screen } from '../../react-testing-library-with-providers';
 import Uploader from '.';
 import {
   uploader,
