+++ src/app/components/Embeds/Uploader/index.tsx
@@ -4,13 +4,13 @@ import { useContext } from 'react';
 import path from 'ramda/src/path';
 import pathOr from 'ramda/src/pathOr';
 import { jsx } from '@emotion/react';
-import { ServiceContext } from '../../contexts/ServiceContext';
-import Paragraph from '../Paragraph';
-import Text from '../Text';
+import { ServiceContext } from '../../../contexts/ServiceContext';
+import Paragraph from '../../Paragraph';
+import Text from '../../Text';
 import styles from './index.styles';
-import { LeftChevron, RightChevron } from '../icons';
-import CallToActionLink from '../CallToActionLink';
-import idSanitiser from '../../lib/utilities/idSanitiser';
+import { LeftChevron, RightChevron } from '../../icons';
+import CallToActionLink from '../../CallToActionLink';
+import idSanitiser from '../../../lib/utilities/idSanitiser';
 
 interface UploaderProps {
   blocks?: object[];
